Generate many separate experiments for engreitz data
We decided to take another approach to this data. Instead of loading 1 combined
experiment we load 30 individual experiments but create a "super experiment"
(which is a feature yet to be added) that contains all 30 and will allow users
to easily look at all the experiments together, but also still look at them individually.

This change basically takes the previous work and makes it so the experiments
aren't all merged. Thankfully most of the work was still usable, as is.