Add support for full data files
Add support for Full compressed data file TSVs. That is, TSVs with rows of

Accession Id\tCompress full data file

Like the file for experiment and analysis metadata.

Also, handle the "no accession ids at all" case, which is now supported on the back end -- the server will
create the next experiment accession automatically if none is passed in.